fix: use `setAttribute` for spellcheck and translate
