+++ .changeset/forty-guests-chew.md
@@ -1,0 +1,5 @@
+---
+'svelte': patch
+---
+
+fix: use `setAttribute` for spellcheck and translate
