+++ packages/svelte/tests/runtime-runes/samples/svelte-element-draggable-spellcheck-translate/main.svelte
@@ -1,0 +1,7 @@
+<script>
+	let attrs = $state({ draggable: 'false', spellcheck: 'false', translate: 'no' });
+</script>
+
+<svelte:element this={'div'} draggable="false" spellcheck="false" translate="no"></svelte:element>
+
+<div {...attrs}></div>
