--- packages/svelte/tests/runtime-runes/samples/svelte-element-draggable/main.svelte
@@ -1,7 +1,0 @@
-<script>
-	let attrs = $state({ draggable: 'false' });
-</script>
-
-<svelte:element this={'div'} draggable="false"></svelte:element>
-
-<div {...attrs}></div>
